# openapi/sendtrap.yaml
openapi: 3.1.0
info:
  title: Sendtrap API
  version: '1.0'
  summary: HTTP API for Sendtrap inboxes — list, inspect, assert on and manage captured email.
  description: |
    Sendtrap is an SMTP catcher / email sandbox. Every inbox exposes this HTTP API,
    authenticated by the inbox's own API token. The API is served by both Sendtrap
    Community (self-hosted) and Sendtrap Cloud from the same core package; this
    document is the authoritative contract for the `/api/v1` surface and the
    Mailtrap-compatible alias routes.

    ## Authentication

    Every request must carry the inbox API token, either as a bearer token:

    ```
    Authorization: Bearer <api_token>
    ```

    or in the `X-Api-Token` header. The token scopes the request to exactly one
    inbox; there is no cross-inbox access. Find the token on the inbox's
    Settings page, or in the `api_token` field of `GET /api/v1/inbox`.

    Requests from IP addresses outside the inbox's configured allow-list are
    rejected with `403` even when the token is valid.

    ## Rate limiting

    Two limiters apply, keyed to the inbox (or to the caller IP when the token
    is invalid):

    - **General** (`inbox-api`): applies to every endpoint. The ceiling is
      host-configured (Community: `SENDTRAP_API_REQUESTS_PER_MINUTE`, unlimited
      by default with a 300/min fallback where a limit applies; Cloud: per plan).
    - **Blocking** (`inbox-api-wait`, 15/min): additionally applies to
...

...

    ## Blocking waits

...

...
servers:
...
security:
  - bearerToken: []
  - apiTokenHeader: []
...
paths:
...
  /v1/messages/{messageId}:
    parameters:
      - $ref: '#/components/parameters/messageId'
    get:
      operationId: getMessage
      tags: [Messages]
      summary: Get a message
      description: |
        Full message detail: parsed bodies, headers in original order,
        extracted links, quality checks, attachment metadata and download URLs.
      responses:
        '200':
          description: The message.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/MessageDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/IpForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    patch:
      operationId: updateMessage
      tags: [Messages]
      summary: Update a message (read state)
      description: The only mutable field is `is_read`. Omitting it marks the message read.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                is_read:
                  type: boolean
                  default: true
      responses:
        '200':
          description: The updated message.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/MessageSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/IpForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    delete:
      operationId: deleteMessage
      tags: [Messages]
      summary: Delete a message
      responses:
        '200':
          description: Message deleted.
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted:
                    type: boolean
                    const: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/IpForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: The inbox API token as a bearer token.
    apiTokenHeader:
      type: apiKey
      in: header
      name: X-Api-Token
      description: The inbox API token as a plain header (alternative to the bearer form).
  parameters:
    messageId:
      name: messageId
      in: path
      required: true
      description: Numeric message id. Messages belonging to a different inbox return `404`.
      schema: { type: integer }
...
  headers:
    X-RateLimit-Limit:
      description: Request ceiling per minute for the applicable limiter.
      schema: { type: integer }
    X-RateLimit-Remaining:
      description: Requests remaining in the current window.
      schema: { type: integer }
  responses:
    Unauthorized:
      description: Missing or invalid API token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing:
              value: { message: 'API token missing.' }
            invalid:
              value: { message: 'Invalid API token.' }
    IpForbidden:
      description: The caller's IP address is outside the inbox's allow-list.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            denied:
              value: { message: 'Access denied for your IP address.' }
    NotFound:
      description: No such resource in this inbox (includes resources owned by other inboxes).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          description: Seconds until the limiter window resets.
          schema: { type: integer }
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      description: Common error envelope for every non-2xx JSON response.
      required: [message]
      properties:
        message:
          type: string
          description: Human-readable explanation of the failure.
    EmailAddress:
      type: object
      description: A parsed mailbox from an address header.
      required: [address, name]
      properties:
        address:
          type: [string, 'null']
        name:
          type: [string, 'null']
    MessageSummary:
      type: object
      description: Lightweight message representation used in lists and assertion results.
      required:
        - id
        - test_id
        - from_address
        - from_name
        - to
        - envelope_to
        - subject
        - size
        - is_read
        - has_attachments
        - has_unresolved_merge_tags
        - received_at
      properties:
        id:
          type: integer
        test_id:
          type: [string, 'null']
          description: Value of the `X-Sendtrap-Test-Id` header, if the sender set one.
        from_address:
          type: [string, 'null']
        from_name:
          type: [string, 'null']
        to:
          type: array
          items:
            $ref: '#/components/schemas/EmailAddress'
          description: Parsed To header recipients.
        envelope_to:
          type: array
          items: { type: string }
          description: SMTP envelope recipients (includes BCC-only recipients).
        subject:
          type: [string, 'null']
        size:
          type: integer
          description: Raw message size in bytes.
        is_read:
          type: boolean
        has_attachments:
          type: boolean
        has_unresolved_merge_tags:
          type: boolean
          description: Whether unreplaced template tags like `{{name}}` were detected.
        received_at:
          type: [string, 'null']
          format: date-time
    MessageDetail:
      type: object
      description: Full message representation with parsed bodies, headers and attachments.
      required:
        - id
        - inbox_id
        - message_id
        - test_id
        - envelope_from
        - envelope_to
        - from_address
        - from_name
        - to
        - cc
        - subject
        - size
        - is_read
        - has_html
        - has_text
        - has_attachments
        - has_unresolved_merge_tags
        - unresolved_merge_tags
        - received_at
        - html
        - text
        - links
        - checks
        - headers
        - attachments
        - urls
      properties:
        id:
          type: integer
        inbox_id:
          type: integer
        message_id:
          type: [string, 'null']
          description: The Message-ID header value, without angle brackets.
        test_id:
          type: [string, 'null']
        envelope_from:
          type: [string, 'null']
          description: SMTP envelope sender (MAIL FROM).
        envelope_to:
          type: array
          items: { type: string }
        from_address:
          type: [string, 'null']
        from_name:
          type: [string, 'null']
        to:
          type: array
          items:
            $ref: '#/components/schemas/EmailAddress'
        cc:
          type: array
          items:
            $ref: '#/components/schemas/EmailAddress'
        subject:
          type: [string, 'null']
        size:
          type: integer
        is_read:
          type: boolean
        has_html:
          type: boolean
        has_text:
          type: boolean
        has_attachments:
          type: boolean
        has_unresolved_merge_tags:
          type: boolean
        unresolved_merge_tags:
          type: array
          items: { type: string }
          description: The unreplaced template tags found in the bodies, e.g. `{{first_name}}`.
        received_at:
          type: [string, 'null']
          format: date-time
        html:
          type: [string, 'null']
          description: Original HTML body (`null` when the message has no HTML part).
        text:
          type: [string, 'null']
          description: Plain-text body.
        links:
          type: array
          items: { type: string }
          description: Deduplicated `href` targets from the HTML body (excluding `mailto:`/`tel:`/fragment-only anchors).
        checks:
          type: array
          items:
            $ref: '#/components/schemas/MessageCheck'
          description: |
            Message-quality lint results. An `html_compatibility` entry is
            included only when a cached HTML Check result exists and the caller
            has HTML Check API access.
        headers:
          type: array
          items:
            type: object
            required: [name, value]
            properties:
              name: { type: string }
              value: { type: string }
          description: All header lines in original order (duplicates preserved).
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/MessageAttachment'
        urls:
          type: object
          required: [raw, html]
          properties:
            raw:
              type: string
              format: uri
            html:
              type: string
              format: uri
    MessageCheck:
      type: object
      required: [key, passed, severity]
      properties:
        key:
          type: string
          description: |
            Check identifier. Current checks: `missing_text_part`,
            `oversized_html`, `missing_list_unsubscribe`,
            `from_address_present`, `html_compatibility`. New checks may be
            added within v1.
        passed:
          type: boolean
        severity:
          type: string
          enum: [info, warn, error]
    MessageAttachment:
      type: object
      required: [id, filename, content_type, size, checksum, is_inline, url]
      properties:
        id:
          type: integer
        filename:
          type: [string, 'null']
        content_type:
          type: [string, 'null']
        size:
          type: integer
        checksum:
          type: [string, 'null']
          description: SHA-256 of the attachment content (hex).
        is_inline:
          type: boolean
          description: Whether the part is an inline (`cid:`-referenced) asset rather than a file attachment.
        url:
          type: string
          format: uri
          description: Authenticated download URL for this attachment.
...